+++ meteor/lib/api/client.ts
@@ -14,6 +14,13 @@ export interface NewClientAPI {
 		functionName: string,
 		...args: any[]
 	): Promise<any>
+	callPeripheralDeviceAction(
+		context: string,
+		deviceId: PeripheralDeviceId,
+		timeoutTime: number | undefined,
+		actionId: string,
+		payload?: Record<string, any>
+	): Promise<any>
 	callBackgroundPeripheralDeviceFunction(
 		deviceId: PeripheralDeviceId,
 		timeoutTime: number | undefined,
@@ -26,6 +33,7 @@ export enum ClientAPIMethods {
 	'clientErrorReport' = 'client.clientErrorReport',
 	'clientLogNotification' = 'client.clientLogNotification',
 	'callPeripheralDeviceFunction' = 'client.callPeripheralDeviceFunction',
+	'callPeripheralDeviceAction' = 'client.callPeripheralDeviceAction',
 	'callBackgroundPeripheralDeviceFunction' = 'client.callBackgroundPeripheralDeviceFunction',
 }