+++ meteor/server/api/peripheralDevice.ts
@@ -4,7 +4,7 @@ import * as _ from 'underscore'
 import { PeripheralDeviceAPI } from '../../lib/api/peripheralDevice'
 import { PeripheralDevices, PeripheralDeviceType, PeripheralDevice } from '../../lib/collections/PeripheralDevices'
 import { Rundowns } from '../../lib/collections/Rundowns'
-import { getCurrentTime, protectString, stringifyObjects, literal } from '../../lib/lib'
+import { getCurrentTime, protectString, stringifyObjects, literal, unprotectString } from '../../lib/lib'
 import { PeripheralDeviceCommands } from '../../lib/collections/PeripheralDeviceCommands'
 import { logger } from '../logging'
 import { TimelineHash } from '../../lib/collections/Timeline'
@@ -63,6 +63,8 @@ import {
 	NewPeripheralDeviceAPI,
 	PeripheralDeviceAPIMethods,
 } from '@sofie-automation/shared-lib/dist/peripheralDevice/methodsAPI'
+import { upsertBundles, generateTranslationBundleOriginId } from './translationsBundles'
+import { isTranslatableMessage } from '@sofie-automation/corelib/dist/TranslatableMessage'
 
 const apmNamespace = 'peripheralDevice'
 export namespace ServerPeripheralDeviceAPI {
@@ -115,7 +117,12 @@ export namespace ServerPeripheralDeviceAPI {
 					parentDeviceId: options.parentDeviceId,
 					versions: options.versions,
 
-					configManifest: options.configManifest,
+					configManifest: options.configManifest
+						? {
+								...options.configManifest,
+								translations: undefined, // unset the translations
+						  }
+						: undefined,
 				},
 				$unset:
 					newVersionsStr !== oldVersionsStr
@@ -150,12 +157,21 @@ export namespace ServerPeripheralDeviceAPI {
 				versions: options.versions,
 				// settings: {},
 
-				configManifest:
-					options.configManifest ??
-					literal<DeviceConfigManifest>({
-						deviceConfig: [],
-					}),
+				configManifest: options.configManifest
+					? {
+							...options.configManifest,
+							translations: undefined,
+					  }
+					: literal<DeviceConfigManifest>({
+							deviceConfig: [],
+					  }),
 			})
+		}
+		if (options.configManifest?.translations) {
+			await upsertBundles(
+				options.configManifest.translations,
+				generateTranslationBundleOriginId(deviceId, 'peripheralDevice')
+			)
 		}
 		return deviceId
 	}
@@ -676,6 +692,13 @@ async function functionReply(
 ): Promise<void> {
 	const device = await checkAccessAndGetPeripheralDevice(deviceId, deviceToken, context)
 
+	if (result && typeof result === 'object' && 'response' in result && isTranslatableMessage(result.response)) {
+		result.response.namespaces = [
+			unprotectString(generateTranslationBundleOriginId(deviceId, 'peripheralDevice')),
+			...(result.response.namespaces || []),
+		]
+	}
+
 	// logger.debug('functionReply', err, result)
 	await PeripheralDeviceCommands.updateAsync(
 		{
@@ -1146,9 +1169,10 @@ class ServerPeripheralDeviceAPIClass extends MethodContextAPI implements NewPeri
 		deviceId: PeripheralDeviceId,
 		deviceToken: string,
 		type: string,
-		packageId: ExpectedPackageId
+		packageId: ExpectedPackageId,
+		removeDelay?: number
 	) {
-		await PackageManagerIntegration.removePackageInfo(this, deviceId, deviceToken, type, packageId)
+		await PackageManagerIntegration.removePackageInfo(this, deviceId, deviceToken, type, packageId, removeDelay)
 	}
 }
 registerClassToMeteorMethods(PeripheralDeviceAPIMethods, ServerPeripheralDeviceAPIClass, false)