+++ scripts/run.js
@@ -80,3 +80,12 @@ function watchMeteor() {
 		}
 	).result;
 })();
+
+function signalHandler(signal) {
+	process.exit();
+}
+
+// Make sure to exit on interupt
+process.on("SIGINT", signalHandler);
+process.on("SIGTERM", signalHandler);
+process.on("SIGQUIT", signalHandler);